+++ sismic/clock/__init__.py
@@ -1,0 +1,3 @@
+from .clock import Clock, SimulatedClock, UtcClock, SynchronizedClock
+
+__all__ = ['Clock', 'SimulatedClock', 'UtcClock', 'SynchronizedClock']