+++ tests/test_interpreter.py
@@ -1,5 +1,4 @@
 import pytest
-
 import pickle
 
 from collections import Counter
@@ -29,14 +28,18 @@ class TestInterpreterWithSimple:
     def test_time(self, interpreter):
         assert interpreter.time == 0
 
-        interpreter.time += 10
+        interpreter.clock.time += 10
+        # No execution means no new time value
+        assert interpreter.time == 0  
+
+        interpreter.execute()
         assert interpreter.time == 10
 
-        interpreter.time = 20
-        assert interpreter.time == 20
-
-        with pytest.raises(ExecutionError):
-            interpreter.time = 10
+    def test_deprecated_time(self, interpreter):
+        with pytest.warns(DeprecationWarning):
+            interpreter.time += 1
+            assert interpreter.time == 0
+            assert interpreter.clock.time == 1
 
     def test_queue(self, interpreter):
         interpreter.queue(Event('e1'))
@@ -532,6 +535,7 @@ def test_run_in_background(elevator):
     task.stop()
 
     assert elevator.context['current'] == 4
+    assert elevator.time == 0
 
 
 class TestCoverageFromTrace: